BillingTable opens ProjectOverview dialog when row clicked
